Extract T8ditor ragdoll editor panel
