+++ T850/T8ditor/CMakeLists.txt
@@ -16,6 +16,7 @@ add_executable(T8ditor
 	EditorWorld.cpp
 	PlayScenePanel.cpp
 	MeshEditorPanel.cpp
+	RagdollEditorPanel.cpp
 	EditorImGui.cpp)
 
 t850_apply_app(T8ditor)
